move errors to error pkg
