+++ usecases/replica/errors/errors.go
@@ -13,6 +13,7 @@ package errors
 
 import (
 	"errors"
+	"fmt"
 )
 
 var (
@@ -25,3 +26,87 @@ var (
 
 	ErrNoDiffFound = errors.New("no diff found")
 )
+
+// StatusCode is communicate the cause of failure during replication
+type StatusCode int
+
+const (
+	StatusOK            = 0
+	StatusClassNotFound = iota + 200
+	StatusShardNotFound
+	StatusNotFound
+	StatusAlreadyExisted
+	StatusNotReady
+	StatusConflict = iota + 300
+	StatusPreconditionFailed
+	StatusReadOnly
+	StatusObjectNotFound
+)
+
+// StatusText returns a text for the status code. It returns the empty
+// string if the code is unknown.
+func StatusText(code StatusCode) string {
+	switch code {
+	case StatusOK:
+		return "ok"
+	case StatusNotFound:
+		return "not found"
+	case StatusClassNotFound:
+		return "class not found"
+	case StatusShardNotFound:
+		return "shard not found"
+	case StatusConflict:
+		return "conflict"
+	case StatusPreconditionFailed:
+		return "precondition failed"
+	case StatusAlreadyExisted:
+		return "already existed"
+	case StatusNotReady:
+		return "local index not ready"
+	case StatusReadOnly:
+		return "read only"
+	case StatusObjectNotFound:
+		return "object not found"
+	default:
+		return ""
+	}
+}
+
+// Error reports error happening during replication
+type Error struct {
+	Code StatusCode `json:"code"`
+	Msg  string     `json:"msg,omitempty"`
+	Err  error      `json:"-"`
+}
+
+// Empty checks whether e is an empty error which equivalent to e == nil
+func (e *Error) Empty() bool {
+	return e.Code == StatusOK && e.Msg == "" && e.Err == nil
+}
+
+// NewError create new replication error
+func NewError(code StatusCode, msg string) *Error {
+	return &Error{code, msg, nil}
+}
+
+func (e *Error) Clone() *Error {
+	return &Error{Code: e.Code, Msg: e.Msg, Err: e.Err}
+}
+
+// Unwrap underlying error
+func (e *Error) Unwrap() error { return e.Err }
+
+func (e *Error) Error() string {
+	return fmt.Sprintf("%s %q: %v", StatusText(e.Code), e.Msg, e.Err)
+}
+
+func (e *Error) IsStatusCode(sc StatusCode) bool {
+	return e.Code == sc
+}
+
+func (e *Error) Timeout() bool {
+	t, ok := e.Err.(interface {
+		Timeout() bool
+	})
+	return ok && t.Timeout()
+}
