+++ usecases/replica/replicator.go
@@ -150,8 +150,8 @@ func (r *Replicator) MergeObject(ctx context.Context,
 	if err != nil {
 		r.log.WithField("op", "merge").WithField("class", r.class).
 			WithField("shard", shard).WithField("uuid", doc.ID).Error(err)
-		var replicaErr *Error
-		if errors.As(err, &replicaErr) && replicaErr != nil && replicaErr.Code == StatusObjectNotFound {
+		var replicaErr *replicaerrors.Error
+		if errors.As(err, &replicaErr) && replicaErr != nil && replicaErr.Code == replicaerrors.StatusObjectNotFound {
 			return objects.NewErrDirtyWriteOfDeletedObject(replicaErr)
 		}
 	}
