+++ usecases/replica/transport.go
@@ -24,6 +24,7 @@ import (
 	"github.com/weaviate/weaviate/entities/search"
 	"github.com/weaviate/weaviate/entities/storobj"
 	"github.com/weaviate/weaviate/usecases/objects"
+	replicaerrors "github.com/weaviate/weaviate/usecases/replica/errors"
 	"github.com/weaviate/weaviate/usecases/replica/hashtree"
 )
 
@@ -39,92 +40,8 @@ type Client interface {
 	WClient
 }
 
-// StatusCode is communicate the cause of failure during replication
-type StatusCode int
-
-const (
-	StatusOK            = 0
-	StatusClassNotFound = iota + 200
-	StatusShardNotFound
-	StatusNotFound
-	StatusAlreadyExisted
-	StatusNotReady
-	StatusConflict = iota + 300
-	StatusPreconditionFailed
-	StatusReadOnly
-	StatusObjectNotFound
-)
-
-// Error reports error happening during replication
-type Error struct {
-	Code StatusCode `json:"code"`
-	Msg  string     `json:"msg,omitempty"`
-	Err  error      `json:"-"`
-}
-
-// Empty checks whether e is an empty error which equivalent to e == nil
-func (e *Error) Empty() bool {
-	return e.Code == StatusOK && e.Msg == "" && e.Err == nil
-}
-
-// NewError create new replication error
-func NewError(code StatusCode, msg string) *Error {
-	return &Error{code, msg, nil}
-}
-
-func (e *Error) Clone() *Error {
-	return &Error{Code: e.Code, Msg: e.Msg, Err: e.Err}
-}
-
-// Unwrap underlying error
-func (e *Error) Unwrap() error { return e.Err }
-
-func (e *Error) Error() string {
-	return fmt.Sprintf("%s %q: %v", StatusText(e.Code), e.Msg, e.Err)
-}
-
-func (e *Error) IsStatusCode(sc StatusCode) bool {
-	return e.Code == sc
-}
-
-// StatusText returns a text for the status code. It returns the empty
-// string if the code is unknown.
-func StatusText(code StatusCode) string {
-	switch code {
-	case StatusOK:
-		return "ok"
-	case StatusNotFound:
-		return "not found"
-	case StatusClassNotFound:
-		return "class not found"
-	case StatusShardNotFound:
-		return "shard not found"
-	case StatusConflict:
-		return "conflict"
-	case StatusPreconditionFailed:
-		return "precondition failed"
-	case StatusAlreadyExisted:
-		return "already existed"
-	case StatusNotReady:
-		return "local index not ready"
-	case StatusReadOnly:
-		return "read only"
-	case StatusObjectNotFound:
-		return "object not found"
-	default:
-		return ""
-	}
-}
-
-func (e *Error) Timeout() bool {
-	t, ok := e.Err.(interface {
-		Timeout() bool
-	})
-	return ok && t.Timeout()
-}
-
 type SimpleResponse struct {
-	Errors []Error `json:"errors,omitempty"`
+	Errors []replicaerrors.Error `json:"errors,omitempty"`
 }
 
 func (r *SimpleResponse) FirstError() error {
@@ -142,8 +59,8 @@ type DeleteBatchResponse struct {
 }
 
 type UUID2Error struct {
-	UUID  string `json:"uuid,omitempty"`
-	Error Error  `json:"error,omitempty"`
+	UUID  string              `json:"uuid,omitempty"`
+	Error replicaerrors.Error `json:"error,omitempty"`
 }
 
 // FirstError returns the first found error
