+++ usecases/replica/transport_test.go
@@ -20,46 +20,46 @@ import (
 	"github.com/pkg/errors"
 	"github.com/stretchr/testify/assert"
 
-	"github.com/weaviate/weaviate/usecases/replica"
+	replicaerrors "github.com/weaviate/weaviate/usecases/replica/errors"
 )
 
 func TestReplicationErrorTimeout(t *testing.T) {
 	ctx := context.Background()
 	ctx, cancel := context.WithDeadline(ctx, time.Now())
 	defer cancel()
-	err := &replica.Error{Err: ctx.Err()}
+	err := &replicaerrors.Error{Err: ctx.Err()}
 	assert.True(t, err.Timeout())
 	err = err.Clone()
 	assert.ErrorIs(t, err, context.DeadlineExceeded)
 }
 
 func TestReplicationErrorMarshal(t *testing.T) {
-	rawErr := replica.Error{Code: replica.StatusClassNotFound, Msg: "Article", Err: errors.New("error cannot be marshalled")}
+	rawErr := replicaerrors.Error{Code: replicaerrors.StatusClassNotFound, Msg: "Article", Err: errors.New("error cannot be marshalled")}
 	bytes, err := json.Marshal(&rawErr)
 	assert.Nil(t, err)
-	got := replica.NewError(0, "")
+	got := replicaerrors.NewError(0, "")
 	assert.Nil(t, json.Unmarshal(bytes, got))
-	want := &replica.Error{Code: replica.StatusClassNotFound, Msg: "Article"}
+	want := &replicaerrors.Error{Code: replicaerrors.StatusClassNotFound, Msg: "Article"}
 	assert.Equal(t, want, got)
 }
 
 func TestReplicationErrorStatus(t *testing.T) {
 	tests := []struct {
-		code replica.StatusCode
+		code replicaerrors.StatusCode
 		desc string
 	}{
 		{-1, ""},
-		{replica.StatusOK, "ok"},
-		{replica.StatusClassNotFound, "class not found"},
-		{replica.StatusShardNotFound, "shard not found"},
-		{replica.StatusNotFound, "not found"},
-		{replica.StatusAlreadyExisted, "already existed"},
-		{replica.StatusConflict, "conflict"},
-		{replica.StatusPreconditionFailed, "precondition failed"},
-		{replica.StatusReadOnly, "read only"},
+		{replicaerrors.StatusOK, "ok"},
+		{replicaerrors.StatusClassNotFound, "class not found"},
+		{replicaerrors.StatusShardNotFound, "shard not found"},
+		{replicaerrors.StatusNotFound, "not found"},
+		{replicaerrors.StatusAlreadyExisted, "already existed"},
+		{replicaerrors.StatusConflict, "conflict"},
+		{replicaerrors.StatusPreconditionFailed, "precondition failed"},
+		{replicaerrors.StatusReadOnly, "read only"},
 	}
 	for _, test := range tests {
-		got := replica.StatusText(test.code)
+		got := replicaerrors.StatusText(test.code)
 		if got != test.desc {
 			t.Errorf("StatusText(%d) want %v got %v", test.code, test.desc, got)
 		}
